[Addon] Cache commonly parsed XML files
During a scan of libraries (Tested using Movies and TV shows) each directory is queried for 2x settings. If no settings are found the parent (recursively) directory settings is used; This commonly defaults to the root directory. During the scan of each sub-directory the same XML file is commonly parsed which is now optimized using a cache.